+++ src/cvv.js
@@ -1,18 +1,51 @@
 var isString = require('lodash.isstring');
 var DEFAULT_LENGTH = 3;
 
+function includes(array, thing) {
+  var i = 0;
+
+  for (; i < array.length; i++) {
+    if (thing === array[i]) { return true; }
+  }
+
+  return false;
+}
+
+function min(array) {
+  var minimum = DEFAULT_LENGTH;
+  var i = 0;
+
+  for (; i < array.length; i++) {
+    minimum = array[i] < minimum ? array[i] : minimum;
+  }
+
+  return minimum;
+}
+
+function max(array) {
+  var maximum = DEFAULT_LENGTH;
+  var i = 0;
+
+  for (; i < array.length; i++) {
+    maximum = array[i] > maximum ? array[i] : maximum;
+  }
+
+  return maximum;
+}
+
 function verification(isValid, isPotentiallyValid) {
   return {isValid: isValid, isPotentiallyValid: isPotentiallyValid};
 }
 
 function cvv(value, maxLength) {
   maxLength = maxLength || DEFAULT_LENGTH;
+  maxLength = maxLength instanceof Array ? maxLength : [maxLength];
 
   if (!isString(value)) { return verification(false, false); }
   if (!/^\d*$/.test(value)) { return verification(false, false); }
-  if (value.length === maxLength) { return verification(true, true); }
-  if (value.length < maxLength) { return verification(false, true); }
-  if (value.length > maxLength) { return verification(false, false); }
+  if (includes(maxLength, value.length)) { return verification(true, true); }
+  if (value.length < min(maxLength)) { return verification(false, true); }
+  if (value.length > max(maxLength)) { return verification(false, false); }
 
   return verification(true, true);
 }